Add Pull Indicator to Header
