+++ js/views/reusable/header.jsx
@@ -21,36 +21,59 @@ class Header extends React.Component {
   render() {
     return (
       <View style={styles.wrapper}>
-        <View style={styles.textWrapper}>
-          <Text style={styles.text} numberOfLines={1}>
-            {this.props.title}
-          </Text>
+        <View style={styles.pillWrapper}>
+          <View style={styles.pill} />
         </View>
-        <TouchableOpacity
-          style={styles.close}
-          onPress={this.props.backFn || this.triggerBack}
-        >
-          <View style={styles.closeIcon}>
-            <CloseIcon style={{ fill: vars.headerColor }} />
+        <View style={styles.bottomWrapper}>
+          <View style={styles.textWrapper}>
+            <Text style={styles.text} numberOfLines={1}>
+              {this.props.title}
+            </Text>
           </View>
-        </TouchableOpacity>
+          <TouchableOpacity
+            style={styles.close}
+            onPress={this.props.backFn || this.triggerBack}
+          >
+            <View style={styles.closeIcon}>
+              <CloseIcon style={{ fill: vars.headerColor }} />
+            </View>
+          </TouchableOpacity>
+        </View>
       </View>
     )
   }
 }
+const paddingVertical = 12
 const styles = StyleSheet.create({
   wrapper: {
     height: vars.headerHeight,
+    backgroundColor: '#fff',
+    paddingBottom: paddingVertical,
+    touchAction: 'none',
+  },
+  pillWrapper: {
+    height: paddingVertical,
+    paddingTop: paddingVertical / 2,
+  },
+  pill: {
+    backgroundColor: '#d8d8d8',
+    width: 36,
+    height: 5,
+    borderRadius: 5,
+    marginLeft: 'auto',
+    marginRight: 'auto',
+  },
+  bottomWrapper: {
     display: 'flex',
     flexDirection: 'row',
-    backgroundColor: '#fff',
   },
   textWrapper: {
     flex: 1,
     paddingLeft: vars.padding,
   },
   text: {
-    lineHeight: vars.headerHeight,
+    fontFamily: vars.fontFamily,
+    lineHeight: vars.headerHeight - paddingVertical * 2,
     color: vars.headerColor,
     fontSize: 18,
     fontWeight: '600',
